requirements.txt
python-telegram-bot==20.7
feedparser==6.0.10
beautifulsoup4==4.12.2
requests==2.31.0
